Code cleanup, remove old duplicate templates, and move sortArguments to new util
