--- src/app/routes/data/slash/resources/resourcesMessage.js
@@ -1,9 +1,0 @@
-const message = {
-      token: process.env.SLACK_TOKEN,
-      as_user: true,
-      link_names: true,
-      mrkdwn_in: ['text', 'pretext'],
-    };
-
-
-module.exports = { message };
